requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
torch_dct==0.1.6
numpy==1.26.4
scipy==1.14.1
soundfile==0.12.1
librosa==0.11.0
matplotlib==3.8.4
hydra-core==1.3.2
omegaconf==2.3.0
python-dotenv==1.0.1
tqdm==4.67.0
wandb==0.18.7
pandas==2.3.3
tabulate==0.9.0
ipython==9.1.0
SciencePlots==2.1.1
ipympl==0.9.7